return to float16

+++ AFQ/tasks/segmentation.py
@@ -99,7 +99,7 @@ def segment(
         raise ValueError("Fatal: No bundles recognized.")
 
     if is_trx:
-        seg_sft.sft.dtype_dict = {"positions": np.float32, "offsets": np.uint32}
+        seg_sft.sft.dtype_dict = {"positions": np.float16, "offsets": np.uint32}
         tgram = TrxFile.from_sft(seg_sft.sft)
         tgram.groups = seg_sft.bundle_idxs
 
@@ -155,7 +155,7 @@ def export_bundles(base_fname, output_dir, bundles, tracking_params):
                 logger.info(f"Saving {fname}")
                 if is_trx:
                     seg_sft.sft.dtype_dict = {
-                        "positions": np.float32,
+                        "positions": np.float16,
                         "offsets": np.uint32,
                     }
                     trxfile = TrxFile.from_sft(bundle_sft)
